Added tests for the new travel time providers based on the shortest paths library. Also changed the strategy for creating query managers on systems with multiple processing cores.
Former-commit-id: 6095541a24ea767777f1973c64c25e47489cf73f

+++ src/test/java/cz/agents/amodsim/ridesharing/vga/mock/TestOnDemandVehicle.java
@@ -47,7 +47,7 @@ public class TestOnDemandVehicle implements MovingEntity{
 	
 
 	@Override
-	public int getVelocity() {
+	public double getVelocity() {
 		return config.vehicleSpeedInMeters;
 	}
